Move <OrderStatus> under MailPoet integration
[MAILPOET-6135]
