+++ mailpoet/assets/js/src/automation/integrations/mailpoet/analytics/components/tabs/orders/rows.tsx
@@ -3,9 +3,9 @@ import { OrderCell } from './cells/order';
 import { CustomerCell } from './cells/customer';
 import { ProductsCell } from './cells/products';
 import { EmailCell } from './cells/email';
+import { StatusCell } from './cells/status';
 import { formattedPrice } from '../../../formatter';
 import { MailPoet } from '../../../../../../../mailpoet';
-import { OrderStatus } from '../../../../../../components/status';
 
 export function transformOrdersToRows(data: OrderSection['data'] | undefined) {
   const orders = data?.items;
@@ -38,7 +38,7 @@ export function transformOrdersToRows(data: OrderSection['data'] | undefined) {
         },
         {
           display: (
-            <OrderStatus
+            <StatusCell
               status={order.details.status.id}
               name={order.details.status.name}
             />
